+++ README.md
@@ -1,0 +1,6 @@
+# Communication subspace
+Implementation of the
+[communication subspace analysis](https://doi.org/10.1016/j.neuron.2019.01.026) using scikit-learn.
+
+Loosely following the
+[MATLAB implementation](https://github.com/joao-semedo/communication-subspace/tree/master)
